expansionリファクタリング - 他ファイルから参照しない関数にstatic - headerとコメント以外のNorm合わせ - echo関数呼び出しコメントアウト
