+++ includes/minishell.h
@@ -189,6 +189,7 @@ int                 expand_redirs(t_node *node);
 int                 tokenize_with_expansion(t_exp_tkn **head, char *str, int env_flag);
 int	                merge_expansion_tokens(t_exp_tkn **head);
 int					split_exp_tokens(t_exp_tkn **head);
+t_exp_tkn	        *extract_literal(char **s, int len, bool is_quoted);
 t_exp_tkn	        *expand_env_var(char **s, bool is_quoted);
 int	update_args_from_exp(t_exp_tkn *head, t_node *node);
 t_exp_tkn	        *new_exp_token(char *str, bool is_expanded, bool is_quoted);
