+++ srcs/execution/exec_builtin.c
@@ -11,6 +11,7 @@ int	exec_builtin_cmd(t_node *node)
 	cmd = node->argv[0];
 	if (ft_strncmp(cmd, "echo", 5) == 0)
 		stat = echo(node->argc, node->argv);
+    // stat = echo(node->argc, node->argv_lst); // こちらを使ってください
 	else if (ft_strncmp(cmd, "cd", 3) == 0)
 		stat = cd(node->argc, node->argv);
 	else if (ft_strncmp(cmd, "pwd", 4) == 0)
